Major refactoring of ProductViewController and CartViewController code. Also, resolved the issue where the app would crash if all items were zeroed out in cart view (the cause was that zeroed out item would be reomved from cart and when table view would ask for a row, the code would return nil since the cart was empty)

+++ Convention/ProductCellDelegate.h
@@ -12,7 +12,7 @@
 @required
 - (void)VoucherChange:(double)price forIndex:(int)idx;
 
-- (void)QtyChange:(double)qty forIndex:(int)idx;
+- (void)QtyChange:(int)qty forIndex:(int)idx;
 
 - (void)QtyTouchForIndex:(int)idx;
